Use size_t instead of int

+++ src/pdfkernel.cpp
@@ -2469,14 +2469,14 @@ wxPdfDocument::PutPatterns()
             Out("/FunctionType 3");
             Out("/Domain [0 1]");
             Out("/Functions [ ", false);
-            for (int id = 0; id <= (int) nStops ; ++id)
+            for (size_t id = 0; id <= nStops; ++id)
             {
-              OutAscii(wxString::Format(wxS("%d 0 R "), idFuncShading+id), false);
+              OutAscii(wxString::Format(wxS("%d 0 R "), idFuncShading + (int) id), false);
             }
             Out("]");
 
             Out("/Bounds [ ", false);
-            for (int id = 1; id <= (int) nStops; ++id)
+            for (size_t id = 1; id <= nStops; ++id)
             {
               OutAscii(wxPdfUtility::Double2String(stops.Item(id).GetPosition(), 3), false);
               Out(" ", false);
@@ -2484,7 +2484,7 @@ wxPdfDocument::PutPatterns()
             Out("]");
 
             Out("/Encode [ ", false);
-            for (int id = 0; id <= (int) nStops; ++id)
+            for (size_t id = 0; id <= nStops; ++id)
             {
               Out("0 1 ", false);
             }
